requirements.txt
torch
transformers
datasets
evaluate
librosa
pydub
pandas
numpy
jiwer
matplotlib
tqdm
pyyaml
soundfile
requests
